Break from card updates
Have been working on the card system a  lot recently and frankly I just need a break. There are some things I still need to do, but for now I'll focus on implementing the user/admin system and comeback to this later. However, the code is solid enough that I feel confident merging back to the main branch